Converting the position of the game elements to used signed ints
This allows us to deal with elements outside of the game board more
easily. This has introduced a panic in the string rendering at the
moment, as the wrapping isn't dealt with.

+++ src/pretty_rendering/apples.rs
@@ -7,15 +7,11 @@ use std::collections::HashSet;
 pub struct Apples<'a> {
     inner: &'a HashSet<crate::Apple>,
     container: Rect,
-    confines: (usize, usize),
+    confines: (i32, i32),
 }
 
 impl<'a> Apples<'a> {
-    pub fn new(
-        inner: &'a HashSet<crate::Apple>,
-        container: Rect,
-        confines: (usize, usize),
-    ) -> Self {
+    pub fn new(inner: &'a HashSet<crate::Apple>, container: Rect, confines: (i32, i32)) -> Self {
         Apples {
             inner,
             container,
